backend/requirements.txt
Django>=5.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
pyjwt>=2.8.0
requests>=2.31.0
strawberry-graphql>=0.219.0
loguru>=0.7.2
web3>=6.15.0
whitenoise>=6.6.0
django-cors-headers>=4.3.1
Faker>=23.0.0
email-validator>=2.1.0
